requirements.txt
SpeechRecognition 
PyAudio
google-generativeai
python-dotenv